Refactor config to remove use of useState, rather subscribe to Apollo cache
