+++ assets/src/application/services/config/types.ts
@@ -44,8 +44,3 @@ export type ConfigDataProps = {
 	siteUrl: SiteUrlProps;
 	timezone: TimezoneProps;
 };
-
-export type Config = {
-	config: ConfigDataProps;
-	setConfig: React.Dispatch<React.SetStateAction<ConfigDataProps>>;
-};
