+++ assets/src/application/services/utilities/money/useMoneyDisplay.ts
@@ -14,7 +14,7 @@ export type MoneyDisplay = {
 };
 
 const useMoneyDisplay = (): MoneyDisplay => {
-	const { config } = useConfig();
+	const config = useConfig();
 	const afterAmount = config.currency.signB4 ? '' : config.currency.sign;
 	const beforeAmount = config.currency.signB4 ? config.currency.sign : '';
 	const formatMoney = formatAmount(config.currency.decimalPlaces);
