+++ assets/src/domain/shared/services/apollo/queries/currentUser/index.ts
@@ -1,22 +1,6 @@
-import gql from 'graphql-tag';
+export * from './queries';
 
-export const GET_CURRENT_USER: any = gql`
-	query GET_CURRENT_USER {
-		viewer {
-			description
-			email
-			firstName
-			id
-			name
-			nicename
-			nickname
-			lastName
-			locale
-			userId
-			username
-		}
-	}
-`;
+export { default as useCurrentUserQueryOptions } from './useCurrentUserQueryOptions';
 
 export { default as useCurrentUser } from './useCurrentUser';
