+++ assets/src/domain/shared/services/apollo/queries/currentUser/queries.ts
@@ -1,0 +1,19 @@
+import gql from 'graphql-tag';
+
+export const GET_CURRENT_USER: any = gql`
+	query GET_CURRENT_USER {
+		viewer {
+			description
+			email
+			firstName
+			id
+			name
+			nicename
+			nickname
+			lastName
+			locale
+			userId
+			username
+		}
+	}
+`;
