+++ assets/src/domain/shared/services/apollo/queries/generalSettings/queries.ts
@@ -1,0 +1,11 @@
+import gql from 'graphql-tag';
+
+export const GET_GENERAL_SETTINGS: any = gql`
+	query GET_GENERAL_SETTINGS {
+		generalSettings {
+			dateFormat
+			timeFormat
+			timezone
+		}
+	}
+`;
