+++ assets/src/domain/shared/services/apollo/queries/generalSettings/useGeneralSettings.ts
@@ -1,24 +1,17 @@
 import pathOr from 'ramda/src/pathOr';
-import { useApolloClient } from '@apollo/react-hooks';
 
-import { GET_GENERAL_SETTINGS } from '.';
-import { ReadQueryOptions } from '../../../../../eventEditor/services/apollo/queries/types';
+import { GET_GENERAL_SETTINGS } from './';
+import { ReadQueryOptions, useCacheQuery } from '../../../../../eventEditor/services/apollo/queries';
 import { GeneralSettings, GeneralSettingsData } from '../../../../../../application/valueObjects/config/types';
 /**
  * A custom react hook for retrieving GeneralSettings
  */
 const useGeneralSettings = (): GeneralSettings => {
-	const client = useApolloClient();
-	let data: GeneralSettingsData;
+	const options: ReadQueryOptions = {
+		query: GET_GENERAL_SETTINGS,
+	};
+	const { data } = useCacheQuery<GeneralSettingsData>(options);
 
-	try {
-		const options: ReadQueryOptions = {
-			query: GET_GENERAL_SETTINGS,
-		};
-		data = client.readQuery(options);
-	} catch (error) {
-		data = null;
-	}
 	return pathOr<GeneralSettings>(null, ['generalSettings'], data);
 };
