+++ assets/src/domain/shared/services/apollo/queries/generalSettings/useGeneralSettingsQueryOptions.ts
@@ -1,0 +1,12 @@
+import { GET_GENERAL_SETTINGS } from './queries';
+import { ReadQueryOptions } from '@edtrServices/apollo';
+
+const useGeneralSettingsQueryOptions = (): ReadQueryOptions => {
+	const options: ReadQueryOptions = {
+		query: GET_GENERAL_SETTINGS,
+	};
+
+	return options;
+};
+
+export default useGeneralSettingsQueryOptions;
